requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26
PyMySQL==1.0.2
SQLAlchemy==1.4.22
Werkzeug==2.0.1
PyJWT==2.1.0
